Brain-Tumor-Classifier/requirements.txt
transformers
torch
torchvision
scikit-learn
pandas
numpy
streamlit
Pillow
accelerate